+++ frontend/src/App.tsx
@@ -18,6 +18,8 @@ import TransactionForm from "./pages/dashboard/transactions/TransactionForm";
 import ReconciliationPage from "./pages/dashboard/reconciliation/ReconciliationPage";
 import PaymentInbox from "./pages/dashboard/payments/PaymentInbox";
 import IntermediaryRemittances from "./pages/dashboard/intermediaries/IntermediaryRemittances";
+import IntermediaryList from "./pages/dashboard/intermediaries/IntermediaryList";
+import IntermediaryDetail from "./pages/dashboard/intermediaries/IntermediaryDetail";
 import AccountPreferencesPage from "./pages/dashboard/settings/AccountPreferencesPage";
 import LandingPage from "./pages/LandingPage";
 import ProtectedRoute from "./components/ProtectedRoute";
@@ -61,7 +63,9 @@ function App() {
               <Route path="transactions" element={<TransactionList />} />
               <Route path="transactions/new" element={<TransactionForm />} />
               <Route path="payments" element={<PaymentInbox />} />
-              <Route path="intermediaries" element={<IntermediaryRemittances />} />
+              <Route path="intermediaries" element={<IntermediaryList />} />
+              <Route path="intermediaries/remittances" element={<IntermediaryRemittances />} />
+              <Route path="intermediaries/:id" element={<IntermediaryDetail />} />
               <Route path="reconciliation" element={<ReconciliationPage />} />
               <Route path="settings" element={<AccountPreferencesPage />} />
               <Route path="dashboard/settings" element={<Navigate to={SETTINGS_PATH} replace />} />